Configure cluster stack environment according to the specified configuration plugin

+++ chop/subutai/src/main/java/org/apache/usergrid/chop/api/store/subutai/SubutaiInstanceManager.java
@@ -21,12 +21,12 @@ package org.apache.usergrid.chop.api.store.subutai;
 
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Collections;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Set;
 import java.util.UUID;
 
-import org.safehaus.subutai.common.host.ContainerHostState;
 import org.safehaus.subutai.core.environment.rest.ContainerJson;
 import org.safehaus.subutai.core.environment.rest.EnvironmentJson;
 import org.slf4j.Logger;
@@ -34,8 +34,8 @@ import org.slf4j.LoggerFactory;
 
 import org.apache.usergrid.chop.spi.InstanceManager;
 import org.apache.usergrid.chop.spi.LaunchResult;
-import org.apache.usergrid.chop.stack.BasicInstance;
-import org.apache.usergrid.chop.stack.BasicInstanceSpec;
+import org.apache.usergrid.chop.stack.BasicCluster;
+import org.apache.usergrid.chop.stack.CoordinatedCluster;
 import org.apache.usergrid.chop.stack.ICoordinatedCluster;
 import org.apache.usergrid.chop.stack.ICoordinatedStack;
 import org.apache.usergrid.chop.stack.Instance;
@@ -81,16 +81,28 @@ public class SubutaiInstanceManager implements InstanceManager
     @Override
     public void terminateInstances( final Collection<String> instanceIds )
     {
-        UUID environmentId =
-                subutaiClient.getEnvironmentIdByInstanceId( UUID.fromString( instanceIds.iterator().next() ) );
+        if ( instanceIds.size() == 0 ) {
+            LOG.info( "No instance found to be terminated" );
+            return;
+        }
+
+        UUID environmentId = subutaiClient
+                .getEnvironmentIdByInstanceId( UUID.fromString( instanceIds.iterator().next() ) );
         for ( String instanceId : instanceIds ) {
             subutaiClient.destroyInstanceByInstanceId( UUID.fromString( instanceId ) );
         }
 
         // Destroy environment if no instance left inside
+        if( environmentId == null ) {
+            LOG.warn( "Could not find the environment that instance(id:{}) belongs to, " +
+                    "and could not check if there is any instance left on the environment. " +
+                    "Therefore not destroying the environment if environment does not have instance left!" );
+        }
         EnvironmentJson environment =
                 subutaiClient.getEnvironmentByEnvironmentId( environmentId );
         if ( environment.getContainers().size() == 0 ) {
+            LOG.info( "Destroying environment {} since no instance left in it.",
+                    environment.getName()+"(" + environment.getId() + ")" );
             subutaiClient.destroyEnvironment( environmentId );
         }
 
@@ -105,26 +117,21 @@ public class SubutaiInstanceManager implements InstanceManager
 
         List<String> instanceIds = new ArrayList<String>( cluster.getSize() );
         Collection<Instance> instances = new ArrayList<Instance>();
+        // Create the environment
         EnvironmentJson environment = subutaiClient.createStackEnvironment( stack );
+
+        if ( environment == null ) {
+            LOG.error( "Could not create environment for {} stack!", stack.getName() );
+            return new SubutaiLaunchResult( cluster.getInstanceSpec(), Collections.EMPTY_LIST );
+        }
+
         Set<ContainerJson> containerHosts = environment.getContainers();
 
         Iterator<ContainerJson> iterator = containerHosts.iterator();
 
         while( iterator.hasNext() ) {
             ContainerJson containerHost = iterator.next();
-            BasicInstanceSpec instanceSpec = new BasicInstanceSpec();
-            instanceSpec.setImageId( containerHost.getTemplateName() );
-
-            String privateIpAddress = containerHost.getIp();
-            String publicIpAddress = containerHost.getIp();
-            String privateDnsName = containerHost.getHostname();
-            String publicDnsName = containerHost.getHostname();
-
-            ContainerHostState containerState;
-            containerState = containerHost.getState();
-            InstanceState instanceState = InstanceState.fromContainerHostState( containerState );
-            Instance instance = new BasicInstance( containerHost.getId().toString(), instanceSpec, instanceState,
-                    privateDnsName, publicDnsName, privateIpAddress, publicIpAddress );
+            Instance instance = SubutaiUtils.getInstanceFromContainer( containerHost );
             instances.add( instance );
         }
 
@@ -141,15 +148,18 @@ public class SubutaiInstanceManager implements InstanceManager
             }
         }
 
-        LOG.info( "Configuring the cluster {}", cluster.getName() );
+        if ( cluster.getConfiguratorPlugin() != null ) {
+            boolean isConfigurationSuccessful = subutaiClient.configureCluster( stack, cluster, instances );
 
-        boolean isConfigurationSuccessful = subutaiClient.configureCluster( stack, cluster );
-
-        if ( isConfigurationSuccessful )
-            LOG.info( "Cluster {} configured ", cluster.getName() );
+            if ( ! isConfigurationSuccessful ) {
+                LOG.error( "Configuration of {} cluster failed with {} plugin! Destroying environment...",
+                        cluster.getName(), cluster.getConfiguratorPlugin() );
+                subutaiClient.destroyEnvironment( environment.getId() );
+                return new SubutaiLaunchResult( cluster.getInstanceSpec(), Collections.EMPTY_LIST );
+            }
+        }
         else {
-            LOG.error( "Configuration of {} cluster failed!", cluster.getName() );
-            return new SubutaiLaunchResult( cluster.getInstanceSpec(), instances );
+            LOG.warn( "No configurator plugin defined in stack.json, skipping configuration of {}", cluster.getName() );
         }
 
         return new SubutaiLaunchResult( cluster.getInstanceSpec(), instances );
@@ -162,22 +172,43 @@ public class SubutaiInstanceManager implements InstanceManager
         LOG.info( "Creating the runner instances on {}", stack.getDataCenter() );
 
         List<String> instanceIds = new ArrayList<String>( stack.getRunnerCount() );
+        Collection<Instance> runnerInstances = subutaiClient.createRunnersOnEnvironment( stack, spec );
+        Iterator<Instance> iterator = runnerInstances.iterator();
+
+        while( iterator.hasNext() ) {
+            Instance instance = iterator.next();
+            runnerInstances.add( instance );
+        }
+
+        if ( runnerInstances.size() == 0 ) {
+            LOG.error( "Runner runnerInstances could not created for {} stack with {} image!", stack.getName(),
+                    spec.getImageId() );
+            return new SubutaiLaunchResult( spec, Collections.EMPTY_LIST );
+
+        }
+        else if ( runnerInstances.size() != stack.getRunnerCount() ) {
+            LOG.error( "%s runner instances created out of %s! Terminating runner instances", runnerInstances.size(), stack.getRunnerCount() );
+
+            while( iterator.hasNext() ) {
+                Instance instance = iterator.next();
+                subutaiClient.destroyInstanceByInstanceId( UUID.fromString( instance.getId() ) );
+            }
+            return new SubutaiLaunchResult( spec, Collections.EMPTY_LIST );
+        }
 
 
         if ( timeout > SLEEP_LENGTH ) {
-            LOG.info( "Waiting for maximum {} msec until all instances are running", timeout );
+            LOG.info( "Waiting for maximum {} msec until all runner instances are running", timeout );
             boolean stateCheck = waitUntil( instanceIds, InstanceState.Running, timeout );
 
             if ( ! stateCheck ) {
-                LOG.warn( "Waiting for instances to get into Running state has timed out" );
+                LOG.warn( "Waiting for runner instances to get into Running state has timed out" );
             }
         }
 
-        LOG.info( "All the runner instances are running" );
+        LOG.info( "All the runner runner instances are running" );
 
-        Collection<Instance> instances = new ArrayList<Instance>( 0 );
-
-        return new SubutaiLaunchResult( spec, instances );
+        return new SubutaiLaunchResult( spec, runnerInstances );
     }
 
 
@@ -185,6 +216,7 @@ public class SubutaiInstanceManager implements InstanceManager
     public Collection<Instance> getClusterInstances( final ICoordinatedStack stack, final ICoordinatedCluster cluster )
     {
         // TODO implement this functionality
+        LOG.warn( "This method is not implemented" );
         return new ArrayList<Instance>( 0 );
     }
 
@@ -193,12 +225,14 @@ public class SubutaiInstanceManager implements InstanceManager
     public Collection<Instance> getRunnerInstances( final ICoordinatedStack stack )
     {
         // TODO implement this functionality
+        LOG.warn( "This method is not implemented" );
         return new ArrayList<Instance>( 0 );
     }
 
 
     public boolean waitUntil ( Collection<String> instanceIds, InstanceState state,  int timeout ) {
         // TODO implement this functionality
+        LOG.warn( "This method is not implemented" );
         return true;
     }
